[SNAP-322] - As a User I want to have a flexibility in settings and specify the mount volume size.
