+++ WebContent/WEB-INF/init-spring-web-config.xml
@@ -16,10 +16,10 @@
     <bean class="com.sungardas.enhancedsnapshots.service.impl.CryptoServiceImpl"/>
 
     <beans profile="dev">
-        <bean class="com.sungardas.init.CredentialsServiceDev"/>
+        <bean class="com.sungardas.init.InitConfigurationServiceDev"/>
     </beans>
 
     <beans profile="prod">
-        <bean class="com.sungardas.init.CredentialsServiceImpl"/>
+        <bean class="com.sungardas.init.InitConfigurationServiceImpl"/>
     </beans>
 </beans>
